preparing for the ACCU on Sea presentation
